Track Part Files
File Overview:

The part files included are '.cdr', '.SLDPRT', and '.SLDASM'. The '.cdr' file type is intended to be used with CorelDRAW software on the Epilog laser engraver available in the DLC. If you don't know how to use the laser engraver, DLC Staff can show you how to do this. The other file types are for use in Solidworks. If any adjustments need to be made, open them in Solidworks and try making the modifications that way. If this gives you difficulty, then these part files can be used as a reference for dimensioning a new part from scratch.

Creating Gears in Solidworks:

Use the gear toolbox to create the internal ring gear and the spur gear. The best way to do this is to find a YouTube video that shows you how to add the toolbox. For reference, you will open Solidworks and find the icon that says "Toolbox". From there you will navigate to the "ANSI Metric" folder, click "Power Transmission", and finally select the type of gear you would like to model. In this case, the group used "Spur gear" and "Internal Spur Gear" (referred to above as the "Internal Ring Gear"). 

Other Notes:
Different CAD software can be used, but Solidworks provided the most straightforward software in order to generate usable .dxf files for use on the Epilog laser engraver. The .dxf file will need to be opened in CorelDRAW. The '.cdr' file is a file that was saved in CorelDRAW and can be opened directly in CorelDRAW to achieve the exact same dimensioning as the fabricated track currently has.